feat(查询组件): 支持下拉组件tag背景颜色自定义
